+++ infrastructure/cdn-in-a-box/traffic_ops/config.sh
@@ -168,7 +168,10 @@ EOF
 
 cat <<-EOF >/opt/traffic_ops/app/conf/production/riak.conf
 {     "user": "$TV_RIAK_USER",
-  "password": "$TV_RIAK_PASSWORD"
+  "password": "$TV_RIAK_PASSWORD",
+  "tlsConfig": {
+    "maxVersion": 770
+  }
 }
 EOF